// MyPharmacy.WebUI/src/models/pharmacy.ts
export type Pharmacy = {
   id?: number | undefined;
   name?: string | undefined;
   address?: string | undefined;
   city?: string | undefined;
   state?: string | undefined;
   zip?: string | undefined;
   prescriptionsFilled?: number | undefined;
   createdDate?: Date | undefined;
   updatedDate?: Date | undefined;
};